bot_discord/requirement.txt
discord.py
python-dotenv~=1.0.1
pandas~=2.2.0
playwright~=1.43.0
fasttext~=0.9.2
httpx~=0.27.0
numpy~=1.26.3
fastapi~=0.111.0
pydantic~=2.7.4